Create dynamics.py
I'm going to start by commiting what I've done so far of the orbital dynamics simulation in 2D. It tracks to any resolution and with a set amount of point, the orbital position in an earth-moon (static) scenario over any time. It saves into separate time-coded files gravity magnitude and gravity components with the n-format (n, seconds). The simulation recognizes a crash with earth and stops (storing some information), full orbits up to a designated time and has a small time deviation that was documented in an .xls file somewhere. The objects tend to float away but I think I can fix that by looking at correction methods for computational approximations.