Moon object
Saving the moon as an object so that in a few I can also make it rotate around the earth. I also created position, acceleration and acceleration components (not in gs like the satellite's files) as .txt files.

I just realized that we should definitely try out using SQL Lite and compare the time it takes to work with it (since maybe writing to the file will be easier and take less memory (one file with maybe 8 files).
I should also add performance metrics: running time compared to total time/resolution (calculations) and  total time.
If this is going to be taking tens of seconds we should also look into adding a sound package for notification.

Do research on how to increase dedicated CPU to running the program.